refactor: cleanup unresolved command execution
